refactor: remove regl render mode
