+++ packages/core/src/services/config/IConfigService.ts
@@ -16,10 +16,6 @@ export interface ISceneConfig extends IRenderConfig {
   // TODO: 场景是否支持 stencil mask
   stencil?: boolean;
   debug?: boolean;
-  /**
-   * Support regl & @antv/g-device-api now
-   */
-  renderer?: 'regl' | 'device';
 }
 
 export interface IGlobalConfigService {
