+++ packages/renderer/src/utils/HashMap.ts
